Configured basic website layout.
Configure the website layout to have two main parts. The main part is
the boxes that will show how much of each item is left as a percentage.
The secondary part to the right of the main part will be where the user
can assign names to food item id numbers so that the names don't have to
be defined in the ESP32 code or hard coded somewhere else.