Added basic WebSocket functionality
Added the ability for the express server to create a websocket
connection with the react website and emit update events whenever a post
request is made. In production, the ESP32 would make the post request,
the server would emit the update event, and the website would receive
the new data for the food item that had a change in percentage
full-ness.