All necessary database functionality should work.
Client can now set new and existing name mappings and get initial name
and percent state. Fixed bug with initializing percent state. For a
client to set a new name mapping, an ESP32 must have already sent a
percent value mapped to a new id. This item will then be given a default
name that will not be stored in the database. Once the client decides to
change that name mapping to be something other than the default, a new
name mapping is set and posted to the database. TODO: fix visual bugs
like the spacing of the input forms changing when there are multiple
percent entries and the changing of the layout of a percent entry when
its name has spaces in it.